+++ catalog/app/containers/Admin/Buckets/Tabulator/index.ts
@@ -1,0 +1,1 @@
+export { default } from './Tabulator'
